gh-579 separated into pieces but load more stopped working
