+++ Multisig/UI/App/MainContentView.swift
@@ -15,7 +15,6 @@ struct MainContentView<Content: View>: View {
     @Environment(\.managedObjectContext)
     var context: CoreDataContext
 
-
     private let content: Content
 
     init(_ content: Content) {
@@ -24,15 +23,9 @@ struct MainContentView<Content: View>: View {
 
     var body: some View {
         NavigationView {
-//            VStack(spacing: 0) {
-//                SafeHeaderView(showsSafeInfo: $showsSafeInfo)
-//                    .frame(height: ScreenMetrics.safeHeaderHeight)
-
             content
                 .environment(\.managedObjectContext, self.context)
                 .navigationBarItems(leading: selectButton, trailing: switchButton)
-//            }
-//            .edgesIgnoringSafeArea(.top)
         }
         .navigationViewStyle(StackNavigationViewStyle())
         .overlay(
