+++ Multisig/UI/Assets/Balances/TokenBalanceCell.swift
@@ -1,32 +1,12 @@
 //
-//  AssetsView.swift
+//  TokenBalanceCell.swift
 //  Multisig
 //
-//  Created by Andrey Scherbovich on 13.05.20.
+//  Created by Dmitry Bespalov on 07.10.20.
 //  Copyright © 2020 Gnosis Ltd. All rights reserved.
 //
 
 import SwiftUI
-
-struct BalancesView: Loadable {
-    @ObservedObject
-    var model: BalancesViewModel
-
-    init(safe: Safe) {
-        self.model = BalancesViewModel(safe: safe)
-    }
-
-    var body: some View {
-         List {
-            ForEach(model.balances) { tokenBalance in
-                TokenBalanceCell(tokenBalance: tokenBalance)
-            }
-        }
-        .onAppear {
-            self.trackEvent(.assetsCoins)
-        }
-    }
-}
 
 struct TokenBalanceCell: View {
     var tokenBalance: TokenBalance
