+++ Multisig/UI/UI Library/NetworkContentView.swift
@@ -1,0 +1,33 @@
+//
+//  NetworkContentView.swift
+//  Multisig
+//
+//  Created by Dmitry Bespalov on 07.10.20.
+//  Copyright © 2020 Gnosis Ltd. All rights reserved.
+//
+
+import SwiftUI
+
+// View that is able to load the content from network
+struct NetworkContentView<Content: View>: View {
+    var status: ViewLoadingStatus
+    var reload: () -> Void = { }
+    var content: () -> Content
+
+    var body: some View {
+        if status == .initial {
+            Text("Loading...").onAppear(perform: reload)
+        } else if status == .loading {
+            FullScreenLoadingView()
+        } else if status == .failure {
+            NoDataView(reload: reload)
+        } else if status == .success {
+            content()
+        }
+    }
+
+}
+
+enum ViewLoadingStatus {
+    case initial, loading, success, failure
+}
