+++ Multisig/UI/UI Library/NoDataView.swift
@@ -1,0 +1,28 @@
+//
+//  NoDataView.swift
+//  Multisig
+//
+//  Created by Dmitry Bespalov on 07.10.20.
+//  Copyright © 2020 Gnosis Ltd. All rights reserved.
+//
+
+import SwiftUI
+
+struct NoDataView: View {
+    var reload: () -> Void = {}
+
+    var body: some View {
+        VStack {
+            ReloadButton(reload: reload)
+                .padding(.top, 20)
+
+            HStack {
+                Image("ico-server-error")
+                Text("Data cannot be loaded").title(.gnoMediumGrey)
+            }
+            .padding(.top, 115)
+
+            Spacer()
+        }
+    }
+}
